requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.21
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
zipp==3.17.0
